backend/requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
aiomysql
cryptography
aerich
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
httpx
pydantic-settings
langchain
langchain-openai
langchain-community
email-validator
bcrypt==4.0.1
